+++ src/index.js
@@ -1,4 +1,5 @@
 import SummerNote from './components/SummerNote'
+import './components/SummerNote.css'
 import * as SummernotePlugin from './components/SummernotePlugin'
 export { SummernotePlugin }
 export default SummerNote